Use App::Cmd and Moose to power the program
This is a huge change in the code that breaks almost all functionality
for the time being, so a new branch will be used.

Only the 'config' action currently half-works.  The test was also
refactored to allow more flexibility.

The idea is for git-home-sync to be part of a CPAN distribution
(Git-HomeSync).  Otherwise, App::FatPacker can be used to make the
binary standalone without a dependence on CPAN libraries.  I also want
to use Dist::Zilla in this process.