Successfully begin using MooseX::App::Cmd
Extending MooseX::App::Cmd::Command in App::Git::HomeSync::Command was a
guess, but it seems to be working as I hoped.

One notable difference is that the dry-run option is accessed via
$self->{'dry-run'}, as opposed to $self->dry_run.  I think this is due
to the command-line options being handled by MooseX::Getopt instead of
Getopt::Long::Descriptive.

I need to get this working:
    $ git-home-sync config --help