Add sophistication and interactivity to "init"
Now the software will assist the user in getting a "master" repository
(to be later specified by --repo-path [to be renamed?]) setup via "git
init --bare".

The setup of the working tree repository in the user's home directory
will be more correct thanks to the additional "git config" commands.

In short, "init" behaves in two ways, depending on whether the
--repo-path option is given or not.

Made use of several fancy CPAN modules in the process.  Revised the test
to account for the new variations in command sequences.  Reorganized and
renamed attributes.